+++ src/Attribute.php
@@ -17,7 +17,7 @@ namespace Webklex\PHPIMAP;
  *
  * @package Webklex\PHPIMAP
  */
-class Attribute {
+class Attribute implements \ArrayAccess {
 
     /** @var string $name */
     protected $name;
@@ -74,6 +74,51 @@ class Attribute {
      */
     public function toArray(){
         return $this->__serialize();
+    }
+
+    /**
+     * Determine if a value exists at an offset.
+     *
+     * @param  mixed  $key
+     * @return bool
+     */
+    public function offsetExists($key) {
+        return array_key_exists($key, $this->values);
+    }
+
+    /**
+     * Get a value at a given offset.
+     *
+     * @param  mixed  $key
+     * @return mixed
+     */
+    public function offsetGet($key) {
+        return $this->values[$key];
+    }
+
+    /**
+     * Set the value at a given offset.
+     *
+     * @param  mixed  $key
+     * @param  mixed  $value
+     * @return void
+     */
+    public function offsetSet($key, $value) {
+        if (is_null($key)) {
+            $this->values[] = $value;
+        } else {
+            $this->values[$key] = $value;
+        }
+    }
+
+    /**
+     * Unset the value at a given offset.
+     *
+     * @param  string  $key
+     * @return void
+     */
+    public function offsetUnset($key) {
+        unset($this->values[$key]);
     }
 
     /**